services/environment-sensor/requirements.txt
Pillow==9.3.0
RPi.GPIO==0.7.1
sensirion-gas-index-algorithm==3.2.2
smbus2==0.4.2
